mv getProxySupported to useProxySupported hook
